+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - JPA: findAll(Sort sort) implementation of OpaDataFilterRepository
 - MongoDB: findAll(Sort sort) implementation of OpaDataFilterMongoRepository
 - Core: Pretty json print log of Partial Request
+- JPA: findAll(Pageable pageable) implementation of OpaDatafilterRepository
 
 ## [0.4.4] - 2020-04-21
 ### Added